B&O: Use autoip instead of DHCP
lwIP supports autoip and DHCP. It also supports a cooperation mode
where DHCP takes priority over autoip. We want autoip, but not DHCP on
the LPC.

mbed-os supports DHCP but not autoip (except in the cooperation
mode). This (somewhat hacky) change replaces DHCP with autoip in
LWIPInterface.cpp when LWIP_FORCE_AUTOIP is enabled.